[FIX] create agent file from client.go has agent
